refactor: split Downloads context into facade + sub-modules
Extract Downloads.History (CRUD, queries, filtering), Downloads.Queue
(initiation, duplicate checking, client routing), and Downloads.Transcoding
(transcode job management) from the 2,178 LOC monolithic downloads.ex.

Downloads.ex becomes a thin facade with 39 defdelegate calls.
No behavior changes, all 38 downloads tests pass.
