+++ lib/mydia/downloads/queue.ex
@@ -1,0 +1,1271 @@
+defmodule Mydia.Downloads.Queue do
+  @moduledoc false
+
+  import Ecto.Query, warn: false
+
+  alias Mydia.Repo
+  alias Mydia.Downloads.Download
+  alias Mydia.Downloads.Client
+  alias Mydia.Downloads.Client.Registry
+  alias Mydia.Downloads.History
+  alias Mydia.Downloads.Structs.DownloadMetadata
+  alias Mydia.Indexers.SearchResult
+  alias Mydia.Indexers.Structs.SearchResultMetadata
+  alias Mydia.Settings
+  alias Mydia.Library.MediaFile
+  alias Mydia.Media.Episode
+  alias Mydia.Events
+  require Logger
+
+  ## Public Functions
+
+  def initiate_download(%SearchResult{} = search_result, opts \\ []) do
+    # Use protocol from search result
+    download_type = search_result.download_protocol
+    Logger.info("Download protocol: #{inspect(download_type)} for #{search_result.title}")
+    Logger.info("Full search_result struct: #{inspect(search_result, limit: :infinity)}")
+
+    opts = Keyword.put(opts, :download_type, download_type)
+
+    with :ok <- check_for_duplicate_download(search_result, opts),
+         {:ok, client_config, client_id, detected_type} <-
+           select_and_add_to_client(search_result, opts),
+         {:ok, download} <-
+           create_download_record_with_retry(search_result, client_config, client_id, opts) do
+      # Use detected type as fallback if protocol wasn't set
+      final_type = download_type || detected_type
+
+      Logger.info(
+        "Final download type: #{inspect(final_type)} (original: #{inspect(download_type)}, detected: #{inspect(detected_type)})"
+      )
+
+      # Track event
+      actor_type = Keyword.get(opts, :actor_type, :system)
+      actor_id = Keyword.get(opts, :actor_id, "downloads_context")
+
+      # Get media_item for context if available (preloaded on download)
+      download_with_media = Repo.preload(download, :media_item)
+
+      Events.download_initiated(download_with_media, actor_type, actor_id,
+        media_item: download_with_media.media_item
+      )
+
+      {:ok, download}
+    else
+      {:error, reason} = error ->
+        Logger.warning("Failed to initiate download: #{inspect(reason)}")
+        error
+    end
+  end
+
+  def cancel_download(%Download{} = download, opts \\ []) do
+    with {:ok, client_config} <- find_client_config(download.download_client),
+         {:ok, adapter} <- get_adapter_for_client(client_config),
+         client_map_config = config_to_map(client_config),
+         :ok <-
+           Client.remove_download(adapter, client_map_config, download.download_client_id, opts),
+         {:ok, _deleted} <- History.delete_download(download) do
+      # Track event
+      actor_type = Keyword.get(opts, :actor_type, :user)
+      actor_id = Keyword.get(opts, :actor_id, "unknown")
+
+      Events.download_cancelled(download, actor_type, actor_id)
+
+      {:ok, download}
+    else
+      {:error, reason} ->
+        Logger.warning("Failed to cancel download: #{inspect(reason)}")
+        {:error, reason}
+    end
+  end
+
+  def pause_download(%Download{} = download, opts \\ []) do
+    with {:ok, client_config} <- find_client_config(download.download_client),
+         {:ok, adapter} <- get_adapter_for_client(client_config),
+         client_map_config = config_to_map(client_config),
+         :ok <- Client.pause_torrent(adapter, client_map_config, download.download_client_id) do
+      # Track event
+      actor_type = Keyword.get(opts, :actor_type, :user)
+      actor_id = Keyword.get(opts, :actor_id, "unknown")
+
+      Events.download_paused(download, actor_type, actor_id)
+
+      History.broadcast_download_update(download.id)
+      {:ok, download}
+    else
+      {:error, reason} ->
+        Logger.warning("Failed to pause download in client: #{inspect(reason)}")
+        {:error, reason}
+    end
+  end
+
+  def resume_download(%Download{} = download, opts \\ []) do
+    with {:ok, client_config} <- find_client_config(download.download_client),
+         {:ok, adapter} <- get_adapter_for_client(client_config),
+         client_map_config = config_to_map(client_config),
+         :ok <- Client.resume_torrent(adapter, client_map_config, download.download_client_id) do
+      # Track event
+      actor_type = Keyword.get(opts, :actor_type, :user)
+      actor_id = Keyword.get(opts, :actor_id, "unknown")
+
+      Events.download_resumed(download, actor_type, actor_id)
+
+      History.broadcast_download_update(download.id)
+      {:ok, download}
+    else
+      {:error, reason} ->
+        Logger.warning("Failed to resume download in client: #{inspect(reason)}")
+        {:error, reason}
+    end
+  end
+
+  def clear_completed(%Download{} = download, opts \\ []) do
+    # Try to remove from client first (ignore errors as may already be removed)
+    case find_client_config(download.download_client) do
+      {:ok, client_config} ->
+        case get_adapter_for_client(client_config) do
+          {:ok, adapter} ->
+            client_map_config = config_to_map(client_config)
+
+            # Attempt to remove from client, but don't fail if it's already gone
+            case Client.remove_download(
+                   adapter,
+                   client_map_config,
+                   download.download_client_id,
+                   opts
+                 ) do
+              :ok ->
+                Logger.info("Removed completed download from client",
+                  download_id: download.id,
+                  client: download.download_client
+                )
+
+              {:error, reason} ->
+                Logger.debug("Could not remove from client (may already be removed)",
+                  download_id: download.id,
+                  reason: inspect(reason)
+                )
+            end
+
+          {:error, _} ->
+            Logger.debug("No adapter found for client", client: download.download_client)
+        end
+
+      {:error, _} ->
+        Logger.debug("Client config not found", client: download.download_client)
+    end
+
+    # Always delete the database record
+    case History.delete_download(download) do
+      {:ok, deleted_download} ->
+        # Track event
+        actor_type = Keyword.get(opts, :actor_type, :user)
+        actor_id = Keyword.get(opts, :actor_id, "unknown")
+
+        Events.download_cleared(download, actor_type, actor_id)
+
+        {:ok, deleted_download}
+
+      error ->
+        error
+    end
+  end
+
+  def clear_all_completed(opts \\ []) do
+    # Get all imported downloads
+    imported_downloads =
+      Download
+      |> where([d], not is_nil(d.imported_at))
+      |> Repo.all()
+
+    results =
+      Enum.map(imported_downloads, fn download ->
+        case clear_completed(download, opts) do
+          {:ok, _} -> :ok
+          {:error, _} -> :error
+        end
+      end)
+
+    success_count = Enum.count(results, &(&1 == :ok))
+    {:ok, success_count}
+  end
+
+  def check_for_duplicate_download(search_result, opts) do
+    media_item_id = Keyword.get(opts, :media_item_id)
+    episode_id = Keyword.get(opts, :episode_id)
+    manual? = Keyword.get(opts, :manual, false)
+
+    # Always check for active downloads to prevent downloading the same thing twice
+    with :ok <- check_for_active_download(search_result, media_item_id, episode_id) do
+      # Skip media file check for manual downloads - the user explicitly chose this release
+      # (they may want to upgrade quality or grab a different version)
+      if manual? do
+        :ok
+      else
+        check_for_existing_media_files(search_result, media_item_id, episode_id)
+      end
+    end
+  end
+
+  def check_for_active_download(search_result, media_item_id, episode_id) do
+    # Query for active downloads (not completed and not failed)
+    base_query =
+      Download
+      |> where([d], is_nil(d.completed_at) and is_nil(d.error_message))
+
+    # Add filters based on what we're downloading
+    query =
+      cond do
+        # For episodes, check if there's an active download for this episode
+        episode_id ->
+          where(base_query, [d], d.episode_id == ^episode_id)
+
+        # For season packs, check if there's an active download for same media_item and season
+        media_item_id &&
+            match?(
+              %SearchResultMetadata{season_pack: true, season_number: _},
+              search_result.metadata
+            ) ->
+          season_number = search_result.metadata.season_number
+
+          base_query
+          |> where([d], d.media_item_id == ^media_item_id)
+          |> where([d], ^Mydia.DB.json_is_true(:metadata, "$.season_pack"))
+          |> where(
+            [d],
+            ^Mydia.DB.json_integer_equals(:metadata, "$.season_number", season_number)
+          )
+
+        # For movies or other media, check if there's an active download for this media_item
+        media_item_id ->
+          where(base_query, [d], d.media_item_id == ^media_item_id)
+
+        # No media association (e.g., music, books, adult libraries)
+        # Check by download_url to prevent downloading the same file twice
+        true ->
+          where(base_query, [d], d.download_url == ^search_result.download_url)
+      end
+
+    active_downloads = Repo.all(query)
+
+    if active_downloads == [] do
+      :ok
+    else
+      # Verify each "active" download still exists in the download client
+      case verify_downloads_in_client(active_downloads) do
+        :all_stale ->
+          # All were stale/orphaned - allow the new download
+          :ok
+
+        :has_active ->
+          season_info =
+            case search_result.metadata do
+              %SearchResultMetadata{season_pack: true, season_number: sn} -> " (season #{sn})"
+              _ -> ""
+            end
+
+          Logger.info("Skipping download - active download already exists#{season_info}",
+            media_item_id: media_item_id,
+            episode_id: episode_id
+          )
+
+          {:error, :duplicate_download}
+      end
+    end
+  end
+
+  def check_for_existing_media_files(search_result, media_item_id, episode_id) do
+    alias Mydia.Media.MediaItem
+
+    cond do
+      # For episodes, check if media files already exist for this episode
+      episode_id ->
+        query = from(f in MediaFile, where: f.episode_id == ^episode_id and is_nil(f.trashed_at))
+
+        if Repo.exists?(query) do
+          Logger.info("Skipping download - media files already exist for episode",
+            episode_id: episode_id
+          )
+
+          {:error, :duplicate_download}
+        else
+          :ok
+        end
+
+      # For season packs, check if any episodes in the season already have media files
+      media_item_id &&
+          match?(
+            %SearchResultMetadata{season_pack: true, season_number: _},
+            search_result.metadata
+          ) ->
+        season_number = search_result.metadata.season_number
+
+        # Get all episodes for this season
+        episodes_query =
+          from(e in Episode,
+            where: e.media_item_id == ^media_item_id and e.season_number == ^season_number,
+            select: e.id
+          )
+
+        episode_ids = Repo.all(episodes_query)
+
+        if episode_ids != [] do
+          # Check if any of these episodes have media files
+          media_files_query =
+            from(f in MediaFile, where: f.episode_id in ^episode_ids and is_nil(f.trashed_at))
+
+          if Repo.exists?(media_files_query) do
+            Logger.info(
+              "Skipping download - media files already exist for some episodes in season",
+              media_item_id: media_item_id,
+              season_number: season_number
+            )
+
+            {:error, :duplicate_download}
+          else
+            :ok
+          end
+        else
+          # No episodes found for this season yet - allow download
+          :ok
+        end
+
+      # For media items (movies or TV shows)
+      media_item_id ->
+        # Get the media item to check its type
+        case Repo.get(MediaItem, media_item_id) do
+          %MediaItem{type: "tv_show"} ->
+            # TV shows can have multiple downloads for different seasons/episodes
+            # Don't block based on existing media files - the user may be downloading
+            # additional seasons or complete series packs
+            Logger.debug(
+              "Allowing download for TV show - TV shows can have multiple season downloads",
+              media_item_id: media_item_id
+            )
+
+            :ok
+
+          %MediaItem{type: "movie"} ->
+            # For movies, check if non-trashed media files already exist
+            query =
+              from(f in MediaFile,
+                where: f.media_item_id == ^media_item_id and is_nil(f.trashed_at)
+              )
+
+            if Repo.exists?(query) do
+              Logger.info("Skipping download - media files already exist for movie",
+                media_item_id: media_item_id
+              )
+
+              {:error, :duplicate_download}
+            else
+              :ok
+            end
+
+          nil ->
+            # Media item not found, allow download (shouldn't happen normally)
+            Logger.warning("Media item not found during duplicate check",
+              media_item_id: media_item_id
+            )
+
+            :ok
+        end
+
+      # No media association, can't check for existing files
+      true ->
+        :ok
+    end
+  end
+
+  # --- Issues Tab Functions ---
+
+  def manually_match_download(%Download{} = download, media_item_id, episode_id \\ nil) do
+    save_path = get_in(download.metadata || %{}, ["save_path"])
+
+    # Keep match_status until import succeeds (import job clears it)
+    attrs = %{
+      media_item_id: media_item_id,
+      episode_id: episode_id
+    }
+
+    case History.update_download(download, attrs) do
+      {:ok, updated} ->
+        %{
+          "download_id" => updated.id,
+          "save_path" => save_path,
+          "cleanup_client" => true,
+          "use_hardlinks" => true,
+          "move_files" => false
+        }
+        |> Mydia.Jobs.MediaImport.new()
+        |> Oban.insert()
+
+        {:ok, updated}
+
+      {:error, _changeset} = error ->
+        error
+    end
+  end
+
+  def refresh_match_suggestions(%Download{} = download) do
+    alias Mydia.Downloads.{TorrentParser, TorrentMatcher}
+
+    suggestions =
+      case TorrentParser.parse(download.title) do
+        {:ok, parsed_info} ->
+          try do
+            TorrentMatcher.find_top_candidates(parsed_info,
+              max_results: 3,
+              monitored_only: false
+            )
+          rescue
+            e ->
+              Logger.warning("Failed to find match candidates: #{inspect(e)}",
+                download_id: download.id
+              )
+
+              []
+          end
+
+        _ ->
+          []
+      end
+
+    current_metadata = download.metadata || %{}
+    updated_metadata = Map.put(current_metadata, "match_suggestions", suggestions)
+
+    History.update_download(download, %{metadata: updated_metadata})
+  end
+
+  def resolve_file_mappings(%Download{} = download, mappings) when is_list(mappings) do
+    # Build target_files for the MediaImport job (mappings always use string keys)
+    target_files =
+      Enum.map(mappings, fn mapping ->
+        %{
+          "path" => mapping["path"],
+          "episode_id" => mapping["episode_id"]
+        }
+      end)
+
+    # Don't clear match_status or unresolved_files metadata here — the import job
+    # clears them on success. This preserves data for retry if import fails.
+    case %{
+           "download_id" => download.id,
+           "target_files" => target_files,
+           "use_hardlinks" => true,
+           "move_files" => false
+         }
+         |> Mydia.Jobs.MediaImport.new()
+         |> Oban.insert() do
+      {:ok, _job} -> {:ok, download}
+      {:error, changeset} -> {:error, changeset}
+    end
+  end
+
+  def dismiss_download(%Download{} = download) do
+    History.delete_download(download)
+  end
+
+  def dismiss_all_cancelled do
+    from(d in Download,
+      where:
+        is_nil(d.match_status) and is_nil(d.imported_at) and
+          (not is_nil(d.error_message) or not is_nil(d.import_failed_at))
+    )
+    |> Repo.delete_all()
+  end
+
+  ## Private Functions - Download Initiation
+
+  defp verify_downloads_in_client(downloads) do
+    has_active =
+      Enum.any?(downloads, fn download ->
+        case verify_single_download_in_client(download) do
+          :active -> true
+          :stale -> false
+        end
+      end)
+
+    if has_active, do: :has_active, else: :all_stale
+  end
+
+  defp verify_single_download_in_client(download) do
+    with {:ok, client_config} <- find_client_config(download.download_client),
+         {:ok, adapter} <- get_adapter_for_client(client_config) do
+      client_map_config = config_to_map(client_config)
+
+      case Client.get_status(adapter, client_map_config, download.download_client_id) do
+        {:ok, _status} ->
+          :active
+
+        {:error, %{type: :not_found}} ->
+          Logger.warning("Active download not found in client, marking as failed",
+            download_id: download.id,
+            title: download.title,
+            client: download.download_client
+          )
+
+          History.mark_download_failed(download, "Torrent no longer exists in download client")
+          :stale
+
+        {:error, reason} ->
+          # Client error (connection issue, etc.) - assume download is still active
+          # to avoid accidentally re-downloading
+          Logger.warning("Could not verify download in client, assuming active",
+            download_id: download.id,
+            reason: inspect(reason)
+          )
+
+          :active
+      end
+    else
+      {:error, _reason} ->
+        # Client config not found - can't verify, assume active
+        :active
+    end
+  end
+
+  # Selects appropriate client and adds the download, with smart fallback if type is detected
+  defp select_and_add_to_client(search_result, opts) do
+    download_type = Keyword.get(opts, :download_type)
+
+    # First, prepare the torrent/nzb input (download file if needed)
+    # Pass the indexer name for authentication
+    with {:ok, torrent_input_result} <-
+           prepare_torrent_input(search_result.download_url, search_result.indexer) do
+      # Extract detected type from the downloaded content
+      detected_type =
+        case torrent_input_result do
+          {:file, _body, type} -> type
+          _ -> nil
+        end
+
+      # Use detected type as fallback if download_type is nil
+      final_download_type = download_type || detected_type
+
+      Logger.info(
+        "File type detection: original=#{inspect(download_type)}, detected=#{inspect(detected_type)}, final=#{inspect(final_download_type)}"
+      )
+
+      # Update opts with the final download type and title
+      opts_with_type =
+        opts
+        |> Keyword.put(:download_type, final_download_type)
+        |> Keyword.put(:title, search_result.title)
+
+      # Now select the appropriate client based on the final type
+      with {:ok, client_config} <- select_download_client(opts_with_type),
+           {:ok, adapter} <- get_adapter_for_client(client_config) do
+        # Extract the actual torrent input (without the type)
+        torrent_input =
+          case torrent_input_result do
+            {:file, body, _type} -> {:file, body}
+            other -> other
+          end
+
+        # Add to the selected client
+        case add_torrent_to_client_with_input(
+               adapter,
+               client_config,
+               torrent_input,
+               opts_with_type
+             ) do
+          {:ok, client_id} ->
+            {:ok, client_config, client_id, final_download_type}
+
+          {:error, _} = error ->
+            error
+        end
+      end
+    end
+  end
+
+  # Version of add_torrent_to_client that accepts pre-downloaded input
+  defp add_torrent_to_client_with_input(adapter, client_config, torrent_input, opts) do
+    client_map_config = config_to_map(client_config)
+    category = Keyword.get(opts, :category, client_config.category)
+    title = Keyword.get(opts, :title)
+
+    torrent_opts =
+      []
+      |> maybe_add_opt(:category, category)
+      |> maybe_add_opt(:title, title)
+
+    case Client.add_torrent(adapter, client_map_config, torrent_input, torrent_opts) do
+      {:ok, client_id} ->
+        {:ok, client_id}
+
+      {:error, error} ->
+        {:error, {:client_error, error}}
+    end
+  end
+
+  defp select_download_client(opts) do
+    client_name = Keyword.get(opts, :client_name)
+    download_type = Keyword.get(opts, :download_type)
+
+    cond do
+      # Use specific client if requested
+      client_name ->
+        case find_client_by_name(client_name) do
+          nil -> {:error, {:client_not_found, client_name}}
+          client -> {:ok, client}
+        end
+
+      # Otherwise select by priority, filtered by download type
+      true ->
+        case select_client_by_priority(download_type) do
+          nil -> {:error, :no_clients_configured}
+          client -> {:ok, client}
+        end
+    end
+  end
+
+  defp find_client_by_name(name) do
+    Settings.list_download_client_configs()
+    |> Enum.find(&(&1.name == name && &1.enabled))
+  end
+
+  defp select_client_by_priority(download_type) do
+    # Torrent clients
+    torrent_clients = [:transmission, :qbittorrent, :rtorrent]
+    # Usenet clients
+    usenet_clients = [:nzbget, :sabnzbd]
+
+    client =
+      Settings.list_download_client_configs()
+      |> Enum.filter(& &1.enabled)
+      |> Enum.filter(fn client ->
+        case download_type do
+          :torrent -> client.type in torrent_clients
+          :nzb -> client.type in usenet_clients
+          # No filter if type unknown
+          _ -> true
+        end
+      end)
+      |> Enum.sort_by(& &1.priority, :asc)
+      |> List.first()
+
+    if client do
+      Logger.info(
+        "Selected download client: #{client.name} (type: #{client.type}, priority: #{client.priority}) for download_type: #{download_type}"
+      )
+    else
+      Logger.warning("No suitable client found for download_type: #{download_type}")
+    end
+
+    client
+  end
+
+  defp get_adapter_for_client(client_config) do
+    case Registry.get_adapter(client_config.type) do
+      {:ok, adapter} ->
+        Logger.info("Using adapter #{inspect(adapter)} for client type #{client_config.type}")
+        {:ok, adapter}
+
+      {:error, _} = error ->
+        error
+    end
+  end
+
+  defp create_download_record(search_result, client_config, client_id, opts) do
+    # Build DownloadMetadata struct from search result
+    metadata_attrs = %{
+      size: search_result.size,
+      seeders: search_result.seeders,
+      leechers: search_result.leechers,
+      quality: search_result.quality,
+      download_protocol: search_result.download_protocol
+    }
+
+    # Add season pack metadata if present
+    metadata_attrs =
+      case search_result.metadata do
+        %SearchResultMetadata{season_pack: true, season_number: season_number} ->
+          Map.merge(metadata_attrs, %{
+            season_pack: true,
+            season_number: season_number
+          })
+
+        _ ->
+          metadata_attrs
+      end
+
+    # Create DownloadMetadata struct and convert to map for database storage
+    metadata = metadata_attrs |> DownloadMetadata.new() |> DownloadMetadata.to_map()
+
+    attrs = %{
+      indexer: search_result.indexer,
+      title: search_result.title,
+      download_url: search_result.download_url,
+      download_client: client_config.name,
+      download_client_id: client_id,
+      media_item_id: Keyword.get(opts, :media_item_id),
+      episode_id: Keyword.get(opts, :episode_id),
+      library_path_id: Keyword.get(opts, :library_path_id),
+      metadata: metadata
+    }
+
+    History.create_download(attrs)
+  end
+
+  defp create_download_record_with_retry(search_result, client_config, client_id, opts) do
+    case create_download_record(search_result, client_config, client_id, opts) do
+      {:ok, download} ->
+        {:ok, download}
+
+      {:error, %Ecto.Changeset{} = changeset}
+      when is_struct(changeset, Ecto.Changeset) ->
+        if has_unique_constraint_error?(changeset, :download_client) do
+          Logger.warning(
+            "Unique constraint on download_client_id, cleaning stale record and retrying",
+            client: client_config.name,
+            client_id: client_id
+          )
+
+          case find_stale_download(client_config.name, client_id) do
+            nil ->
+              {:error, changeset}
+
+            stale_download ->
+              Logger.info("Deleting stale download record",
+                download_id: stale_download.id,
+                title: stale_download.title
+              )
+
+              History.delete_download(stale_download)
+              create_download_record(search_result, client_config, client_id, opts)
+          end
+        else
+          {:error, changeset}
+        end
+
+      error ->
+        error
+    end
+  end
+
+  defp has_unique_constraint_error?(%Ecto.Changeset{} = changeset, field) do
+    Enum.any?(changeset.errors, fn
+      {^field, {_msg, opts}} -> Keyword.get(opts, :constraint) == :unique
+      _ -> false
+    end)
+  end
+
+  defp find_stale_download(client_name, client_id) do
+    Download
+    |> where([d], d.download_client == ^client_name and d.download_client_id == ^client_id)
+    |> Repo.one()
+  end
+
+  defp find_client_config(client_name) do
+    case find_client_by_name(client_name) do
+      nil -> {:error, {:client_not_found, client_name}}
+      client -> {:ok, client}
+    end
+  end
+
+  defp config_to_map(config) do
+    %{
+      type: config.type,
+      host: config.host,
+      port: config.port,
+      use_ssl: config.use_ssl,
+      username: config.username,
+      password: config.password,
+      url_base: config.url_base,
+      api_key: config.api_key,
+      connection_settings: config.connection_settings || %{},
+      options: config.connection_settings || %{}
+    }
+  end
+
+  defp prepare_torrent_input(url, indexer_name) do
+    cond do
+      # Magnet links can be used directly
+      String.starts_with?(url, "magnet:") ->
+        {:ok, {:magnet, url}}
+
+      # For HTTP(S) URLs, download the torrent file content
+      # This avoids redirect issues that download clients can't handle
+      String.starts_with?(url, "http://") or String.starts_with?(url, "https://") ->
+        download_torrent_file(url, indexer_name)
+
+      # Unknown format, try as URL
+      true ->
+        {:ok, {:url, url}}
+    end
+  end
+
+  defp download_torrent_file(url, indexer_name) do
+    Logger.info("Downloading file from URL: #{url}")
+
+    # Get download config for the indexer (cookies and FlareSolverr setting)
+    download_config = get_indexer_download_config(indexer_name)
+
+    if download_config.flaresolverr_enabled do
+      Logger.info("Using FlareSolverr for download from: #{indexer_name}")
+      download_via_flaresolverr(url, download_config.cookies)
+    else
+      download_direct(url, download_config.cookie_header)
+    end
+  end
+
+  # Encodes a URL to ensure special characters in the path are properly escaped.
+  # This handles URLs with spaces, brackets, and other characters that would
+  # otherwise cause Req to fail with :invalid_request_target.
+  # Example: "http://host/path/Movie Title (2008).nzb" becomes
+  #          "http://host/path/Movie%20Title%20%282008%29.nzb"
+  defp encode_url(url) when is_binary(url) do
+    case URI.parse(url) do
+      %URI{path: nil} = uri ->
+        # No path to encode
+        URI.to_string(uri)
+
+      %URI{path: path} = uri ->
+        # Encode only the path portion, preserving the rest
+        # Split path into segments and encode each one
+        encoded_path =
+          path
+          |> String.split("/")
+          |> Enum.map(fn segment ->
+            # URI.encode/2 encodes special characters but preserves already-encoded ones
+            URI.encode(segment, &URI.char_unreserved?/1)
+          end)
+          |> Enum.join("/")
+
+        URI.to_string(%{uri | path: encoded_path})
+    end
+  end
+
+  # Download directly with cookies
+  defp download_direct(url, cookie_header) do
+    if cookie_header != "" do
+      Logger.debug("Using auth cookies for download")
+    end
+
+    # Encode the URL to handle special characters (spaces, brackets, etc.)
+    encoded_url = encode_url(url)
+
+    # First check if the URL redirects to a magnet link
+    # by manually following redirects (Req can't handle magnet: scheme)
+    case follow_to_final_url(encoded_url, cookie_header) do
+      {:ok, {:magnet, magnet_url}} ->
+        Logger.debug("URL redirected to magnet link")
+        {:ok, {:magnet, magnet_url}}
+
+      {:ok, {:http, final_url}} ->
+        # Download the actual torrent file with auth cookies
+        req_opts = if cookie_header != "", do: [headers: [{"cookie", cookie_header}]], else: []
+
+        case Req.get(final_url, req_opts) do
+          {:ok, %{status: 200, body: body}} when is_binary(body) ->
+            Logger.info("Successfully downloaded file (#{byte_size(body)} bytes)")
+
+            Logger.info(
+              "Content preview (first 500 chars): #{inspect(String.slice(body, 0, 500))}"
+            )
+
+            # Check if it looks like an NZB file
+            is_nzb = String.contains?(body, "<?xml") and String.contains?(body, "nzb")
+
+            is_torrent =
+              String.starts_with?(body, "d8:announce") or
+                (byte_size(body) > 11 and :binary.part(body, 0, 11) == "d8:announce")
+
+            # Determine file type
+            detected_type =
+              cond do
+                is_nzb -> :nzb
+                is_torrent -> :torrent
+                true -> nil
+              end
+
+            Logger.info(
+              "File type detection: is_nzb=#{is_nzb}, is_torrent=#{is_torrent}, detected_type=#{inspect(detected_type)}"
+            )
+
+            # If we got HTML content (detected_type=nil), try to extract magnet link
+            if detected_type == nil do
+              case extract_magnet_from_html(body) do
+                {:ok, magnet_url} ->
+                  Logger.info("Extracted magnet link from HTML page")
+                  {:ok, {:magnet, magnet_url}}
+
+                {:error, :no_magnet_found} ->
+                  Logger.error(
+                    "Downloaded HTML content but no magnet link found. Page may require further navigation."
+                  )
+
+                  {:error,
+                   {:download_failed,
+                    "Downloaded page is HTML, not a torrent file. No magnet link found on page."}}
+              end
+            else
+              {:ok, {:file, body, detected_type}}
+            end
+
+          {:ok, %{status: status, body: body}} ->
+            # Log the response body for debugging - it often contains error details
+            body_preview =
+              if is_binary(body) and byte_size(body) > 0 do
+                String.slice(to_string(body), 0, 1000)
+              else
+                "(empty body)"
+              end
+
+            Logger.error(
+              "Failed to download torrent file: HTTP #{status}, response: #{body_preview}"
+            )
+
+            {:error, {:download_failed, "HTTP #{status}: #{body_preview}"}}
+
+          {:error, exception} ->
+            Logger.error("Failed to download torrent file: #{inspect(exception)}")
+            {:error, {:download_failed, "Connection error: #{inspect(exception)}"}}
+        end
+
+      {:error, :too_many_redirects} ->
+        Logger.error("Too many redirects when downloading from: #{url}")
+        {:error, {:download_failed, "Too many redirects (maximum 10)"}}
+
+      {:error, {:redirect_error, message}} ->
+        Logger.error("Redirect error for #{url}: #{message}")
+        {:error, {:download_failed, "Redirect error: #{message}"}}
+
+      {:error, {:http_error, exception}} ->
+        Logger.error("HTTP error when downloading from #{url}: #{inspect(exception)}")
+        {:error, {:download_failed, "Connection failed: #{inspect(exception)}"}}
+
+      {:error, {:unexpected_status, status}} ->
+        Logger.error("Unexpected HTTP status #{status} when downloading from: #{url}")
+        {:error, {:download_failed, "Unexpected HTTP status: #{status}"}}
+
+      {:error, reason} ->
+        Logger.error("Failed to download torrent file from #{url}: #{inspect(reason)}")
+        {:error, {:download_failed, inspect(reason)}}
+    end
+  end
+
+  # Download via FlareSolverr for Cloudflare-protected sites
+  defp download_via_flaresolverr(url, cookies) do
+    alias Mydia.Indexers.FlareSolverr
+
+    if not FlareSolverr.enabled?() do
+      Logger.error("FlareSolverr required but not enabled/configured")
+      {:error, {:download_failed, "FlareSolverr required but not configured"}}
+    else
+      # Pass cookies to FlareSolverr request
+      flaresolverr_opts =
+        if cookies != [] do
+          [cookies: cookies]
+        else
+          []
+        end
+
+      case FlareSolverr.get(url, flaresolverr_opts) do
+        {:ok, response} ->
+          body = response.solution.response
+
+          if is_binary(body) and byte_size(body) > 0 do
+            Logger.info("FlareSolverr downloaded file (#{byte_size(body)} bytes)")
+
+            # Check if response is a magnet link redirect
+            if String.starts_with?(body, "magnet:") do
+              {:ok, {:magnet, String.trim(body)}}
+            else
+              # Detect file type
+              is_nzb = String.contains?(body, "<?xml") and String.contains?(body, "nzb")
+
+              is_torrent =
+                String.starts_with?(body, "d8:announce") or
+                  (byte_size(body) > 11 and :binary.part(body, 0, 11) == "d8:announce")
+
+              detected_type =
+                cond do
+                  is_nzb -> :nzb
+                  is_torrent -> :torrent
+                  true -> nil
+                end
+
+              Logger.info(
+                "FlareSolverr file type detection: detected_type=#{inspect(detected_type)}"
+              )
+
+              # If we got HTML content (detected_type=nil), try to extract magnet link
+              if detected_type == nil do
+                case extract_magnet_from_html(body) do
+                  {:ok, magnet_url} ->
+                    Logger.info("Extracted magnet link from HTML page")
+                    {:ok, {:magnet, magnet_url}}
+
+                  {:error, :no_magnet_found} ->
+                    Logger.error(
+                      "FlareSolverr returned HTML content but no magnet link found. Page may require further navigation."
+                    )
+
+                    {:error,
+                     {:download_failed,
+                      "Downloaded page is HTML, not a torrent file. No magnet link found on page."}}
+                end
+              else
+                {:ok, {:file, body, detected_type}}
+              end
+            end
+          else
+            Logger.error("FlareSolverr returned empty response for: #{url}")
+            {:error, {:download_failed, "Empty response from FlareSolverr"}}
+          end
+
+        {:error, reason} ->
+          Logger.error("FlareSolverr download failed: #{inspect(reason)}")
+          {:error, {:download_failed, "FlareSolverr error: #{inspect(reason)}"}}
+      end
+    end
+  end
+
+  defp follow_to_final_url(url, cookie_header, redirects_remaining \\ 10)
+  defp follow_to_final_url(_url, _cookie_header, 0), do: {:error, :too_many_redirects}
+
+  defp follow_to_final_url(url, cookie_header, redirects_remaining) do
+    # Build request options with cookies if available
+    req_opts =
+      [redirect: false] ++
+        if(cookie_header != "", do: [headers: [{"cookie", cookie_header}]], else: [])
+
+    # Try HEAD request first - use redirect: false to get redirect responses directly
+    # instead of following them, which avoids exception handling
+    case Req.head(url, req_opts) do
+      {:ok, %{status: status} = response} when status in 301..308 ->
+        # This is a redirect response
+        case get_location_header(response.headers) do
+          nil ->
+            Logger.error("Redirect (#{status}) missing Location header for URL: #{url}")
+            {:error, {:redirect_error, "Redirect missing Location header"}}
+
+          location ->
+            if String.starts_with?(location, "magnet:") do
+              {:ok, {:magnet, location}}
+            else
+              # Follow the redirect, encoding the location URL to handle special characters
+              follow_to_final_url(encode_url(location), cookie_header, redirects_remaining - 1)
+            end
+        end
+
+      {:ok, %{status: 200}} ->
+        # No redirect, this is the final URL
+        {:ok, {:http, url}}
+
+      {:ok, %{status: 405}} ->
+        # HEAD not allowed, try GET as fallback
+        follow_to_final_url_with_get(url, cookie_header, redirects_remaining)
+
+      {:ok, %{status: status, body: body}} ->
+        body_preview =
+          if is_binary(body) and byte_size(body) > 0 do
+            String.slice(to_string(body), 0, 500)
+          else
+            "(empty body)"
+          end
+
+        Logger.error(
+          "Unexpected HTTP status #{status} during redirect check for URL: #{url}, response: #{body_preview}"
+        )
+
+        {:error, {:unexpected_status, status}}
+
+      {:ok, %{status: status}} ->
+        Logger.error("Unexpected HTTP status #{status} during redirect check for URL: #{url}")
+        {:error, {:unexpected_status, status}}
+
+      {:error, exception} ->
+        {:error, {:http_error, exception}}
+    end
+  end
+
+  defp follow_to_final_url_with_get(url, cookie_header, redirects_remaining) do
+    # Fallback to GET when HEAD is not allowed
+    # Build request options with cookies if available
+    req_opts =
+      [redirect: false] ++
+        if(cookie_header != "", do: [headers: [{"cookie", cookie_header}]], else: [])
+
+    case Req.get(url, req_opts) do
+      {:ok, %{status: status} = response} when status in 301..308 ->
+        # This is a redirect response
+        case get_location_header(response.headers) do
+          nil ->
+            Logger.error("Redirect (#{status}) missing Location header for URL: #{url}")
+            {:error, {:redirect_error, "Redirect missing Location header"}}
+
+          location ->
+            if String.starts_with?(location, "magnet:") do
+              {:ok, {:magnet, location}}
+            else
+              # Follow the redirect, encoding the location URL to handle special characters
+              follow_to_final_url(encode_url(location), cookie_header, redirects_remaining - 1)
+            end
+        end
+
+      {:ok, %{status: 200}} ->
+        # No redirect, this is the final URL
+        {:ok, {:http, url}}
+
+      {:ok, %{status: status, body: body}} ->
+        body_preview =
+          if is_binary(body) and byte_size(body) > 0 do
+            String.slice(to_string(body), 0, 500)
+          else
+            "(empty body)"
+          end
+
+        Logger.error(
+          "Unexpected HTTP status #{status} during GET redirect check for URL: #{url}, response: #{body_preview}"
+        )
+
+        {:error, {:unexpected_status, status}}
+
+      {:ok, %{status: status}} ->
+        Logger.error("Unexpected HTTP status #{status} during GET redirect check for URL: #{url}")
+
+        {:error, {:unexpected_status, status}}
+
+      {:error, exception} ->
+        {:error, {:http_error, exception}}
+    end
+  end
+
+  defp get_location_header(headers) do
+    Enum.find_value(headers, fn
+      {key, [value | _]} when key in ["location", "Location"] -> value
+      {key, value} when key in ["location", "Location"] and is_binary(value) -> value
+      _ -> nil
+    end)
+  end
+
+  defp maybe_add_opt(opts, _key, nil), do: opts
+  defp maybe_add_opt(opts, key, value), do: Keyword.put(opts, key, value)
+
+  # Get download configuration for an indexer by name
+  # Returns a map with cookies (formatted as header string) and flaresolverr_enabled flag
+  defp get_indexer_download_config(nil) do
+    %{cookie_header: "", cookies: [], flaresolverr_enabled: false}
+  end
+
+  defp get_indexer_download_config(indexer_name) when is_binary(indexer_name) do
+    case Mydia.Indexers.get_cardigann_download_config(indexer_name) do
+      nil ->
+        %{cookie_header: "", cookies: [], flaresolverr_enabled: false}
+
+      config ->
+        %{
+          cookie_header: format_cookies_for_header(config.cookies),
+          cookies: config.cookies,
+          flaresolverr_enabled: config.flaresolverr_enabled
+        }
+    end
+  end
+
+  # Convert cookie list to header string
+  # Handles both map format (from FlareSolverr) and string format
+  defp format_cookies_for_header([]), do: ""
+
+  defp format_cookies_for_header(cookies) when is_list(cookies) do
+    cookies
+    |> Enum.map(&format_single_cookie/1)
+    |> Enum.filter(&(&1 != nil))
+    |> Enum.join("; ")
+  end
+
+  defp format_single_cookie(%{"name" => name, "value" => value}) when is_binary(name) do
+    "#{name}=#{value}"
+  end
+
+  defp format_single_cookie(%{name: name, value: value}) when is_binary(name) do
+    "#{name}=#{value}"
+  end
+
+  defp format_single_cookie(cookie) when is_binary(cookie) do
+    # Already a string like "name=value" or "name=value; path=/"
+    # Extract just the name=value part
+    cookie |> String.split(";") |> List.first() |> String.trim()
+  end
+
+  defp format_single_cookie(_), do: nil
+
+  # Extracts a magnet link from HTML content
+  # This is used when FlareSolverr returns an HTML page (e.g., 1337x torrent detail page)
+  # instead of a torrent file
+  defp extract_magnet_from_html(html) when is_binary(html) do
+    # Parse HTML and look for magnet links
+    case Floki.parse_document(html) do
+      {:ok, document} ->
+        # Try multiple strategies to find magnet links
+
+        # Strategy 1: Look for anchor tags with href starting with "magnet:"
+        magnet_links =
+          document
+          |> Floki.find("a[href^='magnet:']")
+          |> Floki.attribute("href")
+
+        # Strategy 2: Also check for data attributes or onclick handlers that might contain magnet
+        magnet_from_data =
+          if magnet_links == [] do
+            document
+            |> Floki.find("[data-href^='magnet:'], [data-url^='magnet:']")
+            |> Floki.attribute("data-href")
+            |> Kernel.++(
+              document
+              |> Floki.find("[data-href^='magnet:'], [data-url^='magnet:']")
+              |> Floki.attribute("data-url")
+            )
+          else
+            []
+          end
+
+        all_magnets = magnet_links ++ magnet_from_data
+
+        # Strategy 3: Regex fallback - look for magnet links in raw HTML
+        all_magnets =
+          if all_magnets == [] do
+            case Regex.scan(~r/magnet:\?xt=urn:[a-zA-Z0-9]+:[a-zA-Z0-9]+[^"'\s<>]*/, html) do
+              [] -> []
+              matches -> Enum.map(matches, fn [match] -> match end)
+            end
+          else
+            all_magnets
+          end
+
+        case all_magnets do
+          [magnet | _] ->
+            # Clean up the magnet link (decode HTML entities)
+            cleaned_magnet =
+              magnet
+              |> String.replace("&amp;", "&")
+              |> String.trim()
+
+            {:ok, cleaned_magnet}
+
+          [] ->
+            {:error, :no_magnet_found}
+        end
+
+      {:error, _reason} ->
+        # Try regex fallback if Floki can't parse the HTML
+        case Regex.run(~r/magnet:\?xt=urn:[a-zA-Z0-9]+:[a-zA-Z0-9]+[^"'\s<>]*/, html) do
+          [magnet | _] ->
+            cleaned_magnet =
+              magnet
+              |> String.replace("&amp;", "&")
+              |> String.trim()
+
+            {:ok, cleaned_magnet}
+
+          nil ->
+            {:error, :no_magnet_found}
+        end
+    end
+  end
+
+  defp extract_magnet_from_html(_), do: {:error, :no_magnet_found}
+end
